fix(gui): 11 参 extern 桥根修——三层 long 对齐(shim long 形参/CTron I64 声明/解释桥 int 栈槽),阶梯 27/27
根因(ldb 四级定位):解释桥 long×N fn-pointer cast 写 8 字节栈槽,int×11 shim 按
4 字节步长读栈参——参 9 后全部错位(hval 读到 u[8] 槽高半=0,bg 读到 u[9]=10)。
≤8 参走寄存器不受累(现存 d_* 钩子全 ≤3 参零风险)。

修复(三层 long 对齐):
- shim gui_cfg/gui_cfg2 形参 int→long(impl 入口 int 截断)
- std/gui.ct gui_cfg 外部声明 I32→I64
- 旧形态夹具(s3/s4/s5/s6/s10-s13/s18)gui_cfg 声明同步 I64
- 解释桥 case≥9 cast int×N + u[] int 化:栈参 4 字节步长写,被调 int 形参逐位对齐
- sem compat 数值宽化:I32 实参可传 I64 形参(已先行落库)

验证:阶梯 27/27;suite 73/73;e8_corpus 15/15;args11 判据 hval=10 bg=11 全对;
s23 发射口径骨架直通全绿;skint2 解释口径组合(when/each/input/Box 闭包键入)全绿。

+++ std/gui/c_src/ctron_gui.c
@@ -147,16 +147,20 @@ static int gui_cfg_impl(int dir, int gap, int padx, int pady, int ax, int ay,
     return 0;
 }
 
-int gui_cfg(int dir, int gap, int padx, int pady, int ax, int ay,
-            int wmode, int wval, int hmode, int hval, int bg_packed) {
-    return gui_cfg_impl(dir, gap, padx, pady, ax, ay, wmode, wval, hmode, hval, bg_packed, 0, 0);
+// 形参用 long:与解释桥 long×N 调用及 CTron 侧 I64 声明对齐(≥9 参走栈参,
+// int 形参 4 字节步长与桥的 8 字节槽错位——args11 实证 hval=0 bg=10)
+int gui_cfg(long dir, long gap, long padx, long pady, long ax, long ay,
+            long wmode, long wval, long hmode, long hval, long bg_packed) {
+    return gui_cfg_impl((int)dir, (int)gap, (int)padx, (int)pady, (int)ax, (int)ay,
+                        (int)wmode, (int)wval, (int)hmode, (int)hval, (int)bg_packed, 0, 0);
 }
 
 // 滚动能力(gui_cfg2):clipv = 垂直裁剪/滚动容器(§13 T0 scroll);
 // offsetpx = 运行时本地滚动偏移(§12.2;滚轮事件由应用侧累计,Clay 按偏移裁剪)
-int gui_cfg2(int dir, int gap, int padx, int pady, int ax, int ay,
-             int wmode, int wval, int hmode, int hval, int bg_packed, int clipv, int offsetpx) {
-    return gui_cfg_impl(dir, gap, padx, pady, ax, ay, wmode, wval, hmode, hval, bg_packed, clipv, offsetpx);
+int gui_cfg2(long dir, long gap, long padx, long pady, long ax, long ay,
+             long wmode, long wval, long hmode, long hval, long bg_packed, long clipv, long offsetpx) {
+    return gui_cfg_impl((int)dir, (int)gap, (int)padx, (int)pady, (int)ax, (int)ay,
+                        (int)wmode, (int)wval, (int)hmode, (int)hval, (int)bg_packed, (int)clipv, (int)offsetpx);
 }
 
 int gui_text(const char *s, int size, int r, int g, int b, int a) {
